test(chat): split provider status panel coverage
Summary:
- move the Claude and Codex provider-status panel assertions into a dedicated
  Project AI panel provider status test file
- shrink `project-conversation-panel-session.test.ts` back under the repo
  max-lines limit while keeping the live transcript coverage in place

Rationale:
- after merging the latest `origin/main`, the session panel test exceeded the
  enforced 300 line limit
- splitting the provider-status assertions preserves coverage and restores web
  lint compatibility for the files touched by this feature

Tests:
- PATH=$HOME/.nvm/versions/node/v22.22.1/bin:$PATH corepack pnpm -C web exec vitest run src/lib/api/chat.test.ts src/lib/features/chat/project-conversation-stream.test.ts src/lib/features/chat/project-conversation-stream-interrupted.test.ts src/lib/features/chat/project-conversation-controller.test.ts src/lib/features/chat/project-conversation-panel.test.ts src/lib/features/chat/project-conversation-panel-session.test.ts src/lib/features/chat/project-conversation-panel-provider-status.test.ts src/lib/features/chat/project-conversation-panel-stop.test.ts src/lib/features/chat/project-conversation-panel-actions.test.ts src/lib/features/chat/project-conversation-controller-actions.test.ts
